Normalize API structure with a template
API_SCAFFOLD is the base hash with each key pointing to a `nil` value.
When merged with the API generated via calling `#hash_for_api`, the
values of the base API will get updated. Therefore, the first two keys
of the final api hash from calling `#to_api` will always be "name" and
"path".

The url fields ("http_url", "api_url", and collections' "entries_url")
are injected towards the end so that they are always the last keys of
the api hash.