Import site metadata into admin interface (#543)
To allow selecting existing categories, layouts and tags via the interface.

The idea is to fetch the *data* related to above every time a page / draft / document is created / edited so that the `Metafields` component always reflects the latest status (i.e. a category added to one post will be available for the next post being edited).